+++ projects/contenteditable/src/lib/contenteditable.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ContenteditableDirective } from './contenteditable.directive';
-
-@NgModule({
-  declarations: [ContenteditableDirective],
-  exports: [ContenteditableDirective],
-})
-export class ContenteditableModule {}