+++ projects/contenteditable/src/lib/editable.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { EditableDirective } from './editable.directive';
+
+@NgModule({
+  declarations: [EditableDirective],
+  exports: [EditableDirective],
+})
+export class ContenteditableModule {}